+++ templates/Choice.html
@@ -3,6 +3,7 @@
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1" />
   <title>Choose Role</title>
+  {%block extra_styles %}
   <style>
     body {
       margin: 0;
@@ -57,6 +58,7 @@
     }
 
     .btn-employer {
+    
       background: linear-gradient(135deg, #00FFB3, #007A5E);
     }
     .btn-employer:hover {
@@ -72,51 +74,6 @@
       transform: scale(1.05);
     }
 
-    .navbar {
-            position: fixed;
-            top: 0;
-            left: 0;
-            right: 0;
-            padding: 1rem 2rem;
-            background: rgba(0, 0, 0, 0.2);
-            backdrop-filter: blur(10px);
-            display: flex;
-            justify-content: space-between;
-            align-items: center;
-            z-index: 1000;
-        }
-
-        .navbar-brand {
-            font-size: 1.5rem;
-            font-weight: 600;
-            color: #fff;
-            text-decoration: none;
-        }
-
-        .nav-links {
-            display: flex;
-            gap: 1.5rem;
-        }
-
-        .nav-link {
-            color: #fff;
-            text-decoration: none;
-            padding: 0.5rem 1rem;
-            border-radius: 8px;
-            transition: all 0.3s ease;
-        }
-
-        .nav-link:hover {
-            background: rgba(255, 255, 255, 0.1);
-        }
-
-        .nav-link.primary {
-            background: #556ee6;
-        }
-
-        .nav-link.primary:hover {
-            background: #3747a8;
-        }
 
 
     @media (max-width: 500px) {
@@ -128,6 +85,7 @@
       }
     }
   </style>
+  {%endblock extra_styles%}
 </head>
 <body>
     {% extends "base.html" %}
