+++ templates/EmployerInformalDisplay.html
@@ -4,6 +4,8 @@
 <meta charset="UTF-8">
 <meta name="viewport" content="width=device-width, initial-scale=1.0">
 <title>Browse Informal Jobs</title>
+{% extends "base.html" %}
+{% block extra_styles %}
 <style>
   @import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;600&display=swap');
 
@@ -146,9 +148,12 @@
     }
   }
 </style>
+{% endblock extra_styles %}
+
+
 </head>
 <body>
-{% extends "base.html" %}
+
 {% block content %}
 <div class="header">
   <a href="/employee/choice" class="back-button">← Back</a>
