+++ templates/base.html
@@ -29,22 +29,28 @@
             padding: 1rem 2rem;
             background: rgba(0, 0, 0, 0.2);
             backdrop-filter: blur(10px);
-            display: flex;
+            display: grid;
             justify-content: space-between;
             align-items: center;
             z-index: 1000;
+            height: 80px;
+                grid-template-columns: 1fr auto 1fr;
         }
 
         .navbar-brand {
-            font-size: 1.5rem;
-            font-weight: 600;
+         
+            
             color: #fff;
             text-decoration: none;
+            position: absolute;
+    left: 50%;
+       top: 50%;                  /* ✅ align vertically */
+    transform: translate(-50%, -50%); 
         }
 
         .nav-links {
             display: flex;
-            gap: 1.5rem;
+            gap: 2 rem;
         }
 
         .nav-link {
@@ -53,6 +59,10 @@
             padding: 0.5rem 1rem;
             border-radius: 8px;
             transition: all 0.3s ease;
+            font-size: 20px;
+            margin-right: 20px;
+            margin-left: 20px;
+            
         }
 
         .nav-link:hover {
@@ -66,7 +76,24 @@
         .nav-link.primary:hover {
             background: #3747a8;
         }
+        
+.navbar-left {
+    justify-self: start;
+    display: flex;
+    align-items: center;
+    gap: 1rem;
+}
 
+.navbar-center {
+    justify-self: center;
+}
+
+.navbar-right {
+    justify-self: end;
+    display: flex;
+    align-items: center;
+    gap: 1rem;
+}
         .main {
             min-height: 100vh;
             display: flex;
@@ -81,8 +108,8 @@
             padding: 3rem;
             border-radius: 20px;
             box-shadow: 0 8px 32px rgba(0, 0, 0, 0.1);
-            max-width: 800px;
-            width: 100%;
+            max-width: 1500px;
+            width: 90%;
             animation: fadeIn 0.5s ease-out;
         }
 
@@ -178,6 +205,7 @@
         h1 {
             font-size: 2rem;
             margin-bottom: 1.5rem;
+            margin-top: 20px;
             color: #fff;
             font-weight: 600;
         }
@@ -192,6 +220,16 @@
             line-height: 1.7;
             color: rgba(255, 255, 255, 0.9);
         }
+        .logo-img {
+               height: 70px;
+    width: 180px;
+    transition: transform 0.3s ease; /* smooth animation */
+}
+
+.logo-img:hover {
+    transform: scale(1.3); /* makes it 10% bigger */
+}
+
 
         @media (max-width: 768px) {
             .navbar {
@@ -230,38 +268,40 @@
              {
                 margin-top: 20px;
             }
-
+            
 
         }
     </style>
     {% block extra_styles %}{% endblock %}
 
-
 <nav class="navbar">
-    <div class = "narbar-left">
-    <a href="/" class="navbar-brand">JobMatcher</a>
-    </div>
+  <div class="navbar-left">
+    {% if session.get('username') %}
+      <a href="/profile" class="nav-link">{{ session.get('username') }}</a>
+    {% endif %}
+  </div>
 
-    <div class="nav-links">
-        {% if session.get('username') %}
-        <div class="navbar-center">
-            <a href = "/profile" class="nav-link" style="margin-right: 20px;"> {{ session.get('username') }}</a>
-        </div>
-   
-           
-            <a href="/" class="nav-link">Home</a>
-             {% block extra_nav %}{% endblock %}
-            <a href="/about" class="nav-link">About Us</a>
-            <a href="/contact" class="nav-link">Contact</a>
-            <a href="#" id="logoutBtn" class="nav-link primary">Logout</a>
-        {% else %}            
-            <a href="/about" class="nav-link">About Us</a>
-            <a href="/contact" class="nav-link">Contact</a>
-            <a href="/login" class="nav-link">Sign In</a>
-        {% endif %}
-     
-    </div>
+  <div class="navbar-center">
+    <a href="/" class="navbar-brand">
+      <img src="static/images/Jinger_logo.png" class="logo-img" >
+    </a>
+  </div>
+
+  <div class="navbar-right">
+    {% if session.get('username') %}
+      <a href="/" class="nav-link">Home</a>
+      {% block extra_nav %}{% endblock %}
+      <a href="/about" class="nav-link">About Us</a>
+      <a href="/contact" class="nav-link">Contact</a>
+      <a href="#" id="logoutBtn" class="nav-link primary">Logout</a>
+    {% else %}
+      <a href="/about" class="nav-link">About Us</a>
+      <a href="/contact" class="nav-link">Contact</a>
+      <a href="/login" class="nav-link">Sign In</a>
+    {% endif %}
+  </div>
 </nav>
+
